+++ src/features/utils.rs
@@ -38,14 +38,14 @@ pub fn collect_metrics(
     report_data: &mut ReportData,
     metadata: std::fs::Metadata,
     path: &std::path::Path,
-    del_meta: (u32, u32),
+    del_meta: (u32, u64),
 ) {
     report_data.files_deleted += del_meta.0;
     report_data.files_scanned += 1;
     report_data.total_file_size_deleted += del_meta.1;
     report_data.total_files_retained = report_data.files_scanned - report_data.files_deleted;
 
-    report_data.total_file_size_retained += metadata.len() as u32;
+    report_data.total_file_size_retained += metadata.len();
     report_data.total_file_size_retained -= del_meta.1;
 
     if let Some(pth) = path.to_str() {
